refactor: split now-playing journal handlers into journal.command.ts
Move the journal sub-feature (~600 lines: header/open/view/page/add/edit/
delete/modal handlers, start-journal, and buildManageJournalButtonRow) into
a focused now-playing/journal.command.ts @Discord class. discordx discovers
handlers across @Discord classes, so registration is unchanged (verified:
identical button/select/modal counts). Journal-component tests updated to
instantiate the new class.

Refs #894
